added number of likes to the like functionality

+++ src/components/Post.js
@@ -16,7 +16,7 @@ import {
   deleteDoc,
 } from "firebase/firestore";
 import { db } from "../../firebase";
-import { useSession } from "next-auth/react";
+import { useSession, signIn } from "next-auth/react";
 import { useState, useEffect } from "react";
 
 export default function Post({ post }) {
@@ -26,12 +26,16 @@ export default function Post({ post }) {
 
   // creates likes collection for the database
   const likePost = async () => {
-    if (hasLiked) {
-      await deleteDoc(doc(db, "posts", post.id, "likes", session?.user.uid));
+    if (session) {
+      if (hasLiked) {
+        await deleteDoc(doc(db, "posts", post.id, "likes", session?.user.uid));
+      } else {
+        await setDoc(doc(db, "posts", post.id, "likes", session.user.uid), {
+          username: session.user.username,
+        });
+      }
     } else {
-      await setDoc(doc(db, "posts", post.id, "likes", session.user.uid), {
-        username: session.user.username,
-      });
+      signIn();
     }
   };
 
@@ -98,17 +102,26 @@ export default function Post({ post }) {
         <div className="flex justify-between text-gray-500 p-2">
           <ChatBubbleOvalLeftEllipsisIcon className="h-9 w-9 hoverEffect p-2 hover:text-sky-500 hover:bg-sky-100  " />
           <TrashIcon className="h-9 w-9 hoverEffect p-2 hover:text-red-600 hover:bg-red-100" />
-          {hasLiked ? (
-            <HeartIconFilled
-              className="h-9 w-9 hoverEffect p-2 text-red-600 hover:bg-red-100"
-              onClick={likePost}
-            />
-          ) : (
-            <HeartIcon
-              className="h-9 w-9 hoverEffect p-2 hover:text-red-600 hover:bg-red-100"
-              onClick={likePost}
-            />
-          )}
+          <div className="flex items-center">
+            {hasLiked ? (
+              <HeartIconFilled
+                className="h-9 w-9 hoverEffect p-2 text-red-600 hover:bg-red-100"
+                onClick={likePost}
+              />
+            ) : (
+              <HeartIcon
+                className="h-9 w-9 hoverEffect p-2 hover:text-red-600 hover:bg-red-100"
+                onClick={likePost}
+              />
+            )}
+            {likes.length > 0 && (
+              <span
+                className={`${hasLiked && "text-red-600"} text-sm select-none`}
+              >
+                {likes.length}
+              </span>
+            )}
+          </div>
           <ShareIcon className="h-9 w-9 hoverEffect p-2 hover:text-sky-500 hover:bg-sky-100" />
           <ChartBarIcon className="h-9 w-9 hoverEffect p-2 hover:text-sky-500 hover:bg-sky-100" />
         </div>
